+++ src/main/java/net/canadensys/dataportal/occurrence/model/OccurrenceViewModel.java
@@ -1,7 +1,11 @@
 package net.canadensys.dataportal.occurrence.model;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+
+import org.apache.commons.lang3.tuple.Pair;
 
 /**
  * This model is tailored for views.
@@ -13,6 +17,8 @@ public class OccurrenceViewModel {
 
 	private List<String> imageList;
 	private List<String> otherMediaList;
+	
+	private Map<String,List<Pair<String,String>>> associatedSequencesPerProviderMap;
 	
 	public void addImage(String image){
 		if(imageList == null){
@@ -28,6 +34,22 @@ public class OccurrenceViewModel {
 		otherMediaList.add(otherMedia);
 	}
 	
+	/**
+	 * @param sequenceProvider
+	 * @param providedSequenceId identifier for the sequence including the provider e.g. GenBank:KC251652
+	 * @param link
+	 */
+	public void addAssociatedSequenceLink(String sequenceProvider, String providedSequenceId, String link){
+		if(associatedSequencesPerProviderMap == null){
+			associatedSequencesPerProviderMap = new HashMap<String, List<Pair<String,String>>>();
+		}
+		
+		if(associatedSequencesPerProviderMap.get(sequenceProvider) == null){
+			associatedSequencesPerProviderMap.put(sequenceProvider, new ArrayList<Pair<String,String>>());
+		}
+		associatedSequencesPerProviderMap.get(sequenceProvider).add(Pair.of(providedSequenceId, link));
+	}
+	
 	public List<String> getImageList() {
 		return imageList;
 	}
@@ -41,5 +63,8 @@ public class OccurrenceViewModel {
 		this.otherMediaList = otherMediaList;
 	}
 	
+	public Map<String,List<Pair<String,String>>> getAssociatedSequencesPerProviderMap() {
+		return associatedSequencesPerProviderMap;
+	}
 	
 }