+++ src/test/resources/insert_test_data.sql
@@ -40,6 +40,21 @@ VALUES (
 'ACAD 3','Myosotis arvensis','species','Canada','Nova Scotia','ACAD','z10441',2013, 'On a lobster claw'
 );
 
+INSERT INTO occurrence(
+auto_id,datasetname,sourcefileid,dwcaid,scientificname,taxonrank,country,stateprovince,collectioncode,catalognumber,habitat,_class,_order,decimallatitude,decimallongitude,associatedSequences)
+VALUES (
+4,'TRT (TRT)','trt-specimens',
+'TRT-1','Acer pseudoplatanus','species','Canada','Nova Scotia','TRT','ECS019597','Under rocks','Magnoliopsida','Sapindales',45.099220,-62.354647,
+'GenBank:KC251652|BOLD:TADCR103-10.rbcLa|GenBank:KC25165x|TestBank:anId'
+);
+INSERT INTO occurrence_raw(
+auto_id,datasetname,sourcefileid,dwcaid,scientificname,taxonrank,country,stateprovince,collectioncode,catalognumber,habitat,_class,_order,decimallatitude,decimallongitude,associatedSequences)
+VALUES (
+4,'TRT (TRT)','trt-specimens',
+'ACAD-1','Acer pseudoplatanus','species','Canada','Nova Scotia','ACAD','ECS019597','Under rocks','Magnoliopsida','Sapindales',45.099220,-62.354647,
+'GenBank:KC251652|BOLD:TADCR103-10.rbcLa|GenBank:KC25165x|TestBank:anId'
+);
+
 INSERT INTO resource_contact(
             id, sourcefileid, resource_name, name, position_name, organization_name, 
             address, city, administrative_area, country, postal_code,