+++ HISTORY.md
@@ -3,6 +3,7 @@ Version History
 
 Version 2.2.0 2014-??-??
 * Controllers are now responsible to provide 'otherLanguage' map
+* Added associatedSequences to occurrence view.
 * Renamed Freemarker root model from 'root' to 'page'
 * New way to deal with web-theme in Gradle